[#462] Add uint8_t

+++ robot/rover/science/include/Carousel.h
@@ -10,11 +10,11 @@
 
 class Carousel : public Stoppable, public Updatable {
     protected:
-        int currentCuvette;
+        uint8_t currentCuvette;
 
     public:
         void home();
-        void goToCuvette(int cuvetteId);
+        void goToCuvette(uint8_t cuvetteId);
         void nextCuvette();
         void previousCuvette();
         int getCurrentCuvette() const { return currentCuvette; }
