+++ robot/rover/science/include/Funnel.h
@@ -11,7 +11,7 @@
 
 class Funnel : public Stoppable, public Updatable {
     protected:
-        int funnelPWM;
+        uint8_t funnelPWM;
     public:
         virtual void funnel();
         virtual void update(unsigned long) override;
